feat: move har-logs logic out from respect-core
