+++ packages/cli/src/__tests__/commands/respect/har-logs/create-har-log.test.ts
@@ -1,4 +1,4 @@
-import { createHarLog } from '../../../utils/har-logs.js';
+import { createHarLog } from '../../../../commands/respect/har-logs/har-logs.js';
 
 describe('createHarLog', () => {
   it('should create a har log', () => {
