+++ packages/respect-core/src/modules/__tests__/flow-runner/context/create-test-context.test.ts
@@ -50,7 +50,6 @@ describe('createTestContext', () => {
     const options = {
       workflowPath: 'modules/description-parser/test.test.yaml',
       workflow: undefined,
-      harLogsFile: 'har-output',
       metadata: {},
       verbose: false,
       maxSteps: 2000,
@@ -69,9 +68,7 @@ describe('createTestContext', () => {
       fetch,
     } as AppOptions;
 
-    const apiClient = new ApiFetcher({
-      harLogs: undefined,
-    });
+    const apiClient = new ApiFetcher({});
     const context = await createTestContext(testDescription, options, apiClient);
 
     expect(context).toMatchObject({
@@ -519,12 +516,10 @@ describe('createTestContext', () => {
       $inputs: {
         env: {},
       },
-      harLogs: expect.any(Object),
       mtlsCerts: undefined,
       options: {
         workflowPath: 'modules/description-parser/test.test.yaml',
         workflow: undefined,
-        harLogsFile: 'har-output',
         metadata: {},
         verbose: false,
       },
@@ -560,7 +555,6 @@ describe('createTestContext', () => {
     const options: AppOptions = {
       workflowPath: 'test.test.yaml',
       workflow: undefined,
-      harOutput: 'har-output',
       metadata: {},
       verbose: false,
       maxSteps: 2000,
@@ -580,9 +574,7 @@ describe('createTestContext', () => {
       fetch,
     };
 
-    const apiClient = new ApiFetcher({
-      harLogs: undefined,
-    });
+    const apiClient = new ApiFetcher({});
     const context = await createTestContext(testDescription, options, apiClient);
 
     expect(context.$workflows.test.inputs).toEqual(undefined);
@@ -616,7 +608,6 @@ describe('createTestContext', () => {
       input: JSON.stringify({ testInput: 'testValue' }),
       workflow: undefined,
       skip: undefined,
-      harOutput: 'har-output',
       metadata: {},
       verbose: false,
       maxSteps: 2000,
@@ -633,9 +624,7 @@ describe('createTestContext', () => {
       fetch,
     };
 
-    const apiClient = new ApiFetcher({
-      harLogs: undefined,
-    });
+    const apiClient = new ApiFetcher({});
     const context = await createTestContext(testDescription, options, apiClient);
 
     expect(context.$workflows).toMatchObject({
@@ -691,9 +680,7 @@ describe('createTestContext', () => {
       input: JSON.stringify({ input1: 'value1' }),
     } as unknown as AppOptions;
 
-    const apiClient = new ApiFetcher({
-      harLogs: undefined,
-    });
+    const apiClient = new ApiFetcher({});
     const context = await createTestContext(testDescription, options, apiClient);
 
     expect(context.$workflows).toEqual({
@@ -734,9 +721,7 @@ describe('createTestContext', () => {
       workflowPath: 'test.test.yaml',
     } as unknown as AppOptions;
 
-    const apiClient = new ApiFetcher({
-      harLogs: undefined,
-    });
+    const apiClient = new ApiFetcher({});
     const context = await createTestContext(testDescription, options, apiClient);
 
     expect(context.$workflows.workflow1.inputs).toEqual(undefined);
@@ -772,9 +757,7 @@ describe('createTestContext', () => {
       },
     } as unknown as AppOptions;
 
-    const apiClient = new ApiFetcher({
-      harLogs: undefined,
-    });
+    const apiClient = new ApiFetcher({});
     const context = await createTestContext(testDescription, options, apiClient);
 
     expect(context.$workflows.workflow1?.inputs?.env?.ENV_VAR).toBe('value');
