+++ packages/respect-core/src/types.ts
@@ -111,8 +111,6 @@ export type AppOptions = {
   workflow?: string | string[];
   skip?: string | string[];
   verbose?: boolean;
-  harOutput?: string;
-  jsonOutput?: string;
   metadata?: Record<string, any>;
   input?: string | string[];
   server?: string | string[];
@@ -243,7 +241,6 @@ export type RunFileResult = {
   totalTimeMs: number;
   totalRequests: number;
   globalTimeoutError: boolean;
-  harLogs: any;
 };
 
 export interface WorkflowExecutionResult {
@@ -277,7 +274,6 @@ export type TestContext = RuntimeExpressionContext & {
   workflows: Workflow[];
   options: AppOptions;
   testDescription: TestDescription;
-  harLogs: any;
   components?: Record<string, any>;
   secretFields: Set<string>;
   severity: Record<string, RuleSeverity>;
